storefront/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
gunicorn
whitenoise
psycopg2-binary
dj-database-url
python-dotenv
djangorestframework-simplejwt